+++ CHANGELOG.md
@@ -427,3 +427,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix custom character card and seen count (kiooeht)
 
 #### dev ####
+* Fix Modal Choice Cards when double-played (kiooeht)