+++ CHANGELOG.md
@@ -429,3 +429,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### dev ####
 * Fix Modal Choice Cards when double-played (kiooeht)
 * Don't shorten name if adding "..." makes it longer than it was (kiooeht)
+* Fix custom character card modifiers (kiooeht)