+++ CHANGELOG.md
@@ -426,9 +426,11 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix ModSlider not allowing x position to be set (kiooeht)
 * Fix custom character card and seen count (kiooeht)
 
-#### dev ####
+#### v4.0.1 ####
 * Fix Modal Choice Cards when double-played (kiooeht)
 * Don't shorten name if adding "..." makes it longer than it was (kiooeht)
 * Fix custom character card modifiers (kiooeht)
 * Allow passing EnergyOrbInterface to CustomPlayer (kiooeht)
 * Slow energy orb spin when out of energy (kiooeht)
+
+#### dev ####